moving common functionalities to common file
